+++ apps/web/src/features/select-profile-image/index.ts
@@ -1,1 +1,2 @@
+export { useProfileImage } from './model/use-profile-image';
 export { default as ProfileImageSheet } from './ui/ProfileImageSheet';
