+++ apps/web/src/views/signup/ui/ProfileSetupStep.tsx
@@ -1,16 +1,18 @@
 'use client';
 
-import { AppBar, Avatar, Button, Field, Icon, Input, Textarea } from '@plog/ui';
+import { useState } from 'react';
 
-import { ProfileImageSheet } from '@/widgets/profile';
+import { useRouter } from 'next/navigation';
+
+import { AppBar } from '@plog/ui';
+
+import { ProfileForm, type ProfileFormData } from '@/widgets/profile-form';
 
 import {
   type DefaultProfileImage,
+  signup,
   type TermsAgreements,
 } from '@/entities/user';
-
-import { useProfileForm } from '../model/use-profile-form';
-import { useProfileImage } from '../model/use-profile-image';
 
 type ProfileSetupStepProps = {
   defaultImages: DefaultProfileImage[];
@@ -23,113 +25,36 @@ export default function ProfileSetupStep({
   termsAgreements,
   onBack,
 }: ProfileSetupStepProps) {
-  const {
-    isSheetOpen,
-    setIsSheetOpen,
-    avatarSrc,
-    imageOption,
-    selectedImageId,
-    setSelectedImageId,
-    openSheet,
-    selectDefault,
-    upload,
-  } = useProfileImage(defaultImages);
+  const router = useRouter();
+  const [isSubmitting, setIsSubmitting] = useState(false);
 
-  const {
+  const handleSubmit = async ({
     nickname,
-    setNickname,
     introduction,
-    setIntroduction,
-    nicknameValidation,
-    introductionValidation,
-    canSubmit,
-    handleSubmit,
-  } = useProfileForm({ termsAgreements, imageOption });
+    imageOption,
+  }: ProfileFormData) => {
+    setIsSubmitting(true);
+    try {
+      await signup(
+        { nickname, introduction: introduction || undefined, termsAgreements },
+        imageOption,
+      );
+      router.push('/');
+    } finally {
+      setIsSubmitting(false);
+    }
+  };
 
   return (
     <>
       <header className="fixed inset-x-0 top-0 z-10 mx-auto max-w-layout">
         <AppBar variant="navigation" title="프로필 설정" onBack={onBack} />
       </header>
-      <form
-        className="flex min-h-dvh w-full flex-col gap-8 p-6 pt-[calc(40px+var(--spacing-header))]"
-        onSubmit={handleSubmit}
-      >
-        <button
-          type="button"
-          aria-label="프로필 이미지 변경"
-          className="relative cursor-pointer self-center"
-          onClick={openSheet}
-        >
-          <Avatar size="large" alt="" src={avatarSrc} />
-          {!avatarSrc && (
-            <Icon
-              name="camera-filled"
-              size={54}
-              className="absolute top-1/2 left-1/2 -translate-1/2 text-semantic-object-normal"
-            />
-          )}
-          <span
-            aria-hidden
-            className="absolute right-0 bottom-0 flex size-11 items-center justify-center rounded-full bg-semantic-accent-normal"
-          >
-            <Icon
-              name="plus"
-              size={30}
-              className="text-semantic-object-inverse"
-            />
-          </span>
-        </button>
-        <Field
-          label="닉네임"
-          required
-          success={
-            nicknameValidation?.status === 'success'
-              ? nicknameValidation.message
-              : undefined
-          }
-          error={
-            nicknameValidation?.status === 'error'
-              ? nicknameValidation.message
-              : undefined
-          }
-        >
-          <Input
-            value={nickname}
-            onClear={() => setNickname('')}
-            onChange={(e) => setNickname(e.target.value)}
-            maxLength={10}
-            placeholder="닉네임을 입력해 주세요."
-          />
-        </Field>
-        <Field
-          className="flex-1"
-          label="소개글"
-          error={
-            introductionValidation?.status === 'error'
-              ? introductionValidation.message
-              : undefined
-          }
-        >
-          <Textarea
-            value={introduction}
-            onChange={(e) => setIntroduction(e.target.value)}
-            maxLength={100}
-            placeholder="개인정보(연락처, SNS 계정 등) 포함 시 노출이 제한될 수 있어요."
-          />
-        </Field>
-        <Button type="submit" size="large" fullWidth disabled={!canSubmit}>
-          시작하기
-        </Button>
-      </form>
-      <ProfileImageSheet
+      <ProfileForm
         defaultImages={defaultImages}
-        open={isSheetOpen}
-        onOpenChange={setIsSheetOpen}
-        selectedImageId={selectedImageId}
-        onSelectedImageIdChange={setSelectedImageId}
-        onSelectDefault={selectDefault}
-        onUpload={upload}
+        submitLabel="시작하기"
+        isSubmitting={isSubmitting}
+        onSubmit={handleSubmit}
       />
     </>
   );
