+++ apps/web/src/widgets/profile-form/index.ts
@@ -1,0 +1,2 @@
+export type { ProfileFormData } from './ui/ProfileForm';
+export { default as ProfileForm } from './ui/ProfileForm';
